Crate lightning_cuda

This crate offers transparent rustic wrappers around the CUDA driver API.

What is this?

This crate offers Rust versions of functions and types from the CUDA driver API. It assumes the user is aware of how the CUDA API looks like, but wants slightly more convenience when working with CUDA.

The goal of this crate is to be as light-weight as possible: most types are a simple wrappers around CUDA types and most functions simply forward the call directly to a CUDA function.

For example, Stream is a newtype around a CUstream, Stream::new simply calls cuStreamCreate, and Drop simply calls cuStreamDestroy, and Stream::raw returns the underlying CUstream when needed for FFI. The should be no runtime overhead of using these methods opposed to the raw functions.

Why choose cuba over other CUDA crates?

There exist many excellent CUDA crates already. The design of this crate is similar to that of RustaCUDA. In fact, CUBA was born out of minor gripes with RustaCUDA’s design, most notably related to FFI interaction, thread-safety, and launching kernels. Although this code in this crate was written from scratch, in some sense, it can be seen as a “spriritual” fork of RustaCUDA.

Some are other crates worth considering when working with CUDA.

Low-level bindings

cuda

cuda-driver-sys

cuda-runtime-sys

cuda-sys

cu-sys

cudart

cu

High-level abstractions

RustaCUDA

accel

coaster

arrayfire

collenchyma

Unsafe and FFI

This crate allows safe functions where possible, but is not afraid to expose unsafe CUDA functionality which cannot be made safe without additional overhead. For example, copy allows copying between buffers if this can be done safely, while the unsafe variant copy_raw performs no safety checks.

Additionally, most types have a raw method which exposes the underlying CUDA type (for example, Stream::raw returns the CUstream) and an unsafe from_raw method which allows constructing a Rust type from a CUDA type (for example, Stream::from_raw which takes a CUstream). This enables FFI integration without existing libraries which work with raw CUDA types.

Error handling

Nearly all functions in the CUDA driver API return an status code indicating whether the operation was successful. This crate mimics this behavior by returning a CudaResult which can be either Ok(_) on success or a CudaError on failure.

Usage

The preferred way of using this crate is importing everything from prelude. The prelude exposes all items prefixed with “Cuda” to prevent name collisions. (for example, Stream becomes CudaStream).

use cuba::prelude::*;

let device = CudaDevice::nth(0)?;
let context = cuda_create_context(device, CudaContextFlags::empty())?;

// Use CudaStream, CudaEvent, CudaModule, CudaDeviceMem, etc...

unsafe { cuda_destroy_context(context)?; }

As an alternative, you could also use the cuba namespace directly, but this is more cluttered.

let device = cuba::Device::nth(0)?;
let context = cuba::create_context(device, cuba::ContextFlags::empty())?;

// Use cuba::Stream, cuba::Event, cuba::Module, cuba::DeviceMem, etc...

unsafe { cuba::destroy_context(context)?; }

Re-exports

pub use context::*;

pub use copy::*;

pub use device::*;

pub use error::*;

pub use event::*;

pub use mem::*;

pub use module::*;

pub use profiler::*;

pub use stream::*;

Modules

context

Management of CUDA contexts

copy

Copying between device/host buffers..

device

CUDA device management.

error

Error handling

event

Management of CUDA events.

mem

Memory allocation in CUDA.

module

Loading CUDA modules and calling kernels.

prelude

Exports commonly used items.

profiler

Management of CUDA profiler.

stream

Management of CUDA streams.

Functions

init

Initialize the CUDA runtime.

version

Returns the version of the CUDA driver API.