Struct lightning_cuda::context::ContextHandle

pub struct ContextHandle(_);

Handle to a CUDA context.

Represents a pointer to valid CUDA context. Use with, try_with, or activate to push this CUDA context on to the thread-local context stack managed by the CUDA runtime.

Note that this types does not own the associated context, but is merely a handle the context. This handle can be obtained using either create_context or retain_device_context. The associated context is not destroyed automatically when the handle goes out of scope, but it must be destroyed manually using either destroy_context or release_device_context when done using it.

Implementations

impl ContextHandle

pub fn current() -> Result<Option<ContextHandle>>

Returns the top CUDA context in the thread-local stack of contexts.

Yields None if the stack is empty.

pub fn device(self) -> Result<Device>

The device associated with this context.

pub fn with<F, T>(self, fun: F) -> Result<T> where
F: FnOnce() -> T,

Call the the given closure while this context is active. This is equivalent to pushing this context, calling the closure, and popping it again.

let context = create_context(Device::nth(0)?, ContextFlags::empty())?;

// Call fun while the context is active.
context.with(|| {
    fun();
})?;

// The above is equivalent to the following, but less error-prone when the work could
// return early.
context.push()?;
fun();
context.pop()?;

pub fn try_with<F, T, E>(self, fun: F) -> Result<T, E> where
F: FnOnce() -> Result<T, E>,
E: From<Error>,

Similar to with, but allows fallible operations.

pub fn activate(&self) -> Result<ContextGuard<'_>>

Make this context the active context. A RAII guard is returned. When the guard is destroyed, the context is deactivated again.

This is equivalent to pushing this context and popping it again once done.

let context = create_context(Device::nth(0)?, ContextFlags::empty())?;

// Call fun while the context is active.
{
    // Must assign guard to variable, otherwise it is immediately destroyed.
    let guard = context.activate();

    // Perform work.
    ..;

    // `guard` goes out of scope and the context is deactivated.
}

// The above is equivalent to the following, but less error-prone when the
// work could return early (e.g., break, return, continue, panic).
{
    context.push()?;
    ..;
    context.pop()?;
}

pub fn push(self) -> Result

Push this context on top of the thread-local stack of context, making it the active context for CUDA operations. The context must be deactivated using pop when done. Usage of this method is not recommended since it is easy to accidentally forget to pop the context. See with, try_with, and activate for alternatives.

pub fn pop() -> Result<Option<ContextHandle>>

Pops the current context of the thread-local stack and returns the handle. Returns None if the stack was empty.

pub fn synchronize(self) -> Result

Blocks until all requested work on this context have completed.

pub fn raw(self) -> CUcontext

Returns the underlying CUcontext of this handle.

pub unsafe fn from_raw(c: CUcontext) -> Option<Self>

Construct a ContextHandle for the given CUcontext. Returns None if the argument is null.

Safety

The given argument must a valid CUcontext as this function performs no additional checks.

pub fn memory_free_and_total(self) -> Result<(usize, usize)>

pub fn enable_peer_access(self) -> Result

Enable peer access

pub fn disable_peer_access(self) -> Result

Disable peer access

pub fn can_access_peer(self, peer: Device) -> Result<bool>