Function lightning_cuda::copy::copy

pub fn copy<S, D, T>(src: S, dst: D) -> Result where
    S: CopySource<Item = T>,
    D: CopyDestination<Item = T>,
    T: Scalar, 

Copies n elements of type T from src to dst.

Parameters src and dst should be memory locations which can be read/written by CUDA (see CopySource and CopyDestination). Type T should be Scalar to ensure the copy operation is safe. The operation is performed synchronously with respect to the host

Panics

Panics if src and dst have different lengths.

Example

// Allocate device memory.
let mut x = DeviceMem::empty(3)?;

// Copy an array.
unsafe {
    copy(&[1, 2, 3], &mut x)?;
}