Function lightning_cuda::copy::copy_raw_async

pub unsafe fn copy_raw_async<T>(
    stream: &Stream, 
    src: DevicePtr<T>, 
    dst: DevicePtr<T>, 
    n: usize
) -> Result

Copies raw bytes asynchronously.

Asynchronously copies n * size_of<T>() raw bytes from src to dst on the provided stream. If possible, prefer to use the safer copy_async function instead.

Safety

src should be a unified virtual address space pointer which should allow for n * size_of<T> consecutive reads.

dst should be a unified virtual address space pointer which should allow for n * size_of<T> consecutive writes.

Elements of type T are bitwise copied which could lead to violating memory safety.

The function returns immediately while the copy is performed asynchronously. The memory locations pointed to by src and dst should be not mutated while the operation is in progress.

Example

// Allocate device memory.
let x = DeviceMem::<i32>::zeroed(10)?;
let mut y = DeviceMem::<i32>::empty(10)?;
let stream = Stream::new()?;

// Copy an array. In this case, we could have also use the regular `copy_async` function.
unsafe {
    copy_raw_async(&stream, x.as_ptr(), y.as_ptr(), 10)?;
}

// Wait until copy completes.
stream.synchronize();