Trait lightning_cuda::copy::CopyDestination

pub unsafe trait CopyDestination {
    type Item;
    fn as_destination(&mut self) -> (DevicePtr<Self::Item>, usize);
}

Memory region suitable to be used as copy destination in copy and copy_async.

Safety

This trait should return a memory region that is valid to be used as destination in cuMemcpy.

Associated Types

type Item

Required methods

fn as_destination(&mut self) -> (DevicePtr<Self::Item>, usize)

Should return a (ptr, len) pair indicating the memory region starts at location ptr and is len * size_of<Item> bytes long.