Struct lightning_cuda::mem::DeviceMem

pub struct DeviceMem<T> { /* fields omitted */ }

Memory buffer allocated on a CUDA device.

Fixed-size buffer of elements of type T located on device memory. The buffer is allocated using cuMemAlloc and freed using cuMemFree.

See Contiguous and ContiguousMut for memory operations on this type.

Implementations

impl<T: Scalar> DeviceMem<T>

pub fn empty(len: usize) -> Result<Self>

Allocate buffer of length len without initializing the memory.

This is safe since any bit pattern is valid for Scalar types.

pub fn zeroed(len: usize) -> Result<Self>

Allocate buffer of length len and set all elements to zero.

This is safe since any bit pattern is valid for Scalar types.

pub fn filled(len: usize, value: T) -> Result<Self>

Allocate buffer of length len and set all elements to value.

This is safe since raw copying elements is valid for Scalar types.

pub fn from_slice(array: &[T]) -> Result<Self>

This is safe since raw copying elements is valid for Scalar types.

impl<T> DeviceMem<T>

pub unsafe fn uninitialized(len: usize) -> Result<Self>

Allocate buffer of length len without initializing the memory.

Safety

The buffer must be initialized before reading the contents.

pub fn into_raw(self) -> (DevicePtr<T>, usize)

Returns the underlying DevicePtr<T> and length of this buffer and consume this object, preventing the destructor from being run.

pub unsafe fn from_raw(ptr: DevicePtr<T>, len: usize) -> Self

Create a DeviceMem<T> from a DevicePtr<T> and length.

Safety

The buffer must have been allocated using cuMemAlloc.

The capacity must be the correct size for the allocation.