Trait lightning_cuda::mem::ContiguousMut

pub trait ContiguousMut<'a>: Contiguous<'a> {
    fn as_slice_mut(&mut self) -> DeviceSliceMut<'a, Self::Item>;

    fn slice_mut(
        &mut self, 
        range: impl SliceRange
    ) -> DeviceSliceMut<'a, Self::Item> { ... }

    fn try_slice_mut(
        &mut self, 
        range: impl SliceRange
    ) -> Option<DeviceSliceMut<'a, Self::Item>> { ... }

    fn split_at_mut(
        &mut self, 
        index: usize
    ) -> (DeviceSliceMut<'a, Self::Item>, DeviceSliceMut<'a, Self::Item>) { ... }

    fn copy_from<S>(&mut self, other: S) -> Result
    where
        S: CopySource<Item = Self::Item>,
        Self::Item: Scalar,
    { ... }

    fn copy_from_slice(&mut self, other: &[Self::Item]) -> Result
    where
        Self::Item: Scalar,
    { ... }

    fn zero(&mut self) -> Result
    where
        Self::Item: Scalar,
    { ... }

    fn fill(&mut self, value: Self::Item) -> Result
    where
        Self::Item: Scalar,
    { ... }

    unsafe fn memset(&mut self, value: u8) -> Result { ... }
}

Required methods

fn as_slice_mut(&mut self) -> DeviceSliceMut<'a, Self::Item>

Provided methods

fn slice_mut(
&mut self,
range: impl SliceRange
) -> DeviceSliceMut<'a, Self::Item>

Returns a

Panics

if the given range is out of bounds.

// Allocate an array of 150 integer.
let mut mem: DeviceMem<u32> = DeviceMem::empty(150)?;

// Create slice to first 50 integers and set their value to 1.
mem.slice_mut(..50).fill(1);

// Create slice to middle 50 integers and set their value to 0.
mem.slice_mut(50..100).zero();

fn try_slice_mut(
&mut self,
range: impl SliceRange
) -> Option<DeviceSliceMut<'a, Self::Item>>

Similar to slice_mut, expect it returns None if the range is out of bounds instead of panicking.

fn split_at_mut(
&mut self,
index: usize
) -> (DeviceSliceMut<'a, Self::Item>, DeviceSliceMut<'a, Self::Item>)

fn copy_from<S>(&mut self, other: S) -> Result where
S: CopySource<Item = Self::Item>,
Self::Item: Scalar,

Copy data from another buffer into this buffer. Alias for copy(other, self).

fn copy_from_slice(&mut self, other: &[Self::Item]) -> Result where
Self::Item: Scalar,

Copy data from a slice into this buffer. Alias for copy(other, self).

fn zero(&mut self) -> Result where
Self::Item: Scalar,

Set all elements of this buffer to zero. This is safe since any bit pattern is valid for Scalar types.

fn fill(&mut self, value: Self::Item) -> Result where
Self::Item: Scalar,

Set all elements in this buffer to value. This is safe since raw copying elements is valid for Scalar types.

unsafe fn memset(&mut self, value: u8) -> Result

Set all bytes in this memory region to the given byte value.

Safety

The caller must ensure that writing raw bytes results in a valid value for T.