Trait lightning_cuda::mem::Scalar

pub unsafe trait Scalar: Copy + Send + Sync + Sized + 'static { }

Represents a type which can be safely copied back-and-forth between host and device memory without leading to memory safety violations.

Safety

The type must be Copy (No Drop implementation).

The type must Send and Sync (Possible to copy between threads via device memory).

Any possible bit-pattern must be valid value. This allows methods such as Device::empty to exist which allocate memory without initializing. It also means copy data from device memory to host memory is always safe even for faulty kernels which corrupt memory. This restriction is the reason why bool and char are not Scalar.