Struct lightning_cuda::stream::Stream

#[repr(transparent)]
pub struct Stream(_);

CUDA Compute stream.

CUDA uses the concept of streams to organize asynchronous operations. A stream is basically a queue of work performed on host and the device (e.g., kernel launches, memory copies). This object wraps a CUstream.

Implementations

impl Stream

pub fn new() -> Result<Self>

Construct a new stream.

pub fn with_options(priority: Option<i32>, flags: StreamFlags) -> Result<Self>

Construct a new stream with the given priority and flags. The given priority should be in the range provided by Stream::priority_range(). Lower numbers represent higher priorities.

pub fn default() -> Result<Self>

Returns the CUDA default stream.

pub fn is_default(&self) -> bool

Returns true if this stream represents the CUDA default stream.

pub fn priority_range() -> Result<(i32, i32)>

Returns the numerical range for stream priorities as a tuple of (least-priority, greatest-priority). Note that lower priority number represent higher priorities, i.e., least-priority >= greatest-priority.

pub fn priority(&self) -> Result<i32>

Returns the priority of this stream.

pub fn wait_for_event(&self, event: &Event) -> Result

Makes all future work for this stream wait until the given event has fired.

pub fn add_callback<F>(&self, fun: F) where
F: FnOnce(Result) + Send + 'static,

Add a callback onto this stream. The callback will be called after previously enqueued work has finished and will be block work added after it. See cuStreamAddCallback for further restrictions and guarantees.

The callback will be passed either Ok(()) or an error code, either because CUDA failed to enqueue the callback or because an asynchronous device error occurred.

The callback must not panic. Any panic will abort the entire process as to prevent the panic from crossing from Rust into the CUDA runtime.

pub fn query(&self) -> Result<bool>

Returns true if all previously enqueued work onto the stream has completed and false otherwise.

pub fn synchronize(&self) -> Result

Block until all previously enqueued work onto this stream has completed.

pub fn raw(&self) -> CUstream

Returns the underlying CUstream object.

pub fn into_raw(self) -> CUstream

Returns the underlying CUstream object and consume this object, preventing the destructor from being run.

pub unsafe fn from_raw(stream: CUstream) -> Self

Construct from a CUstream object.

Safety

The given CUstream object should be a valid CUDA stream object.